pulumi/mcp-server/requirements.txt
# MCP Server Dependencies
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database backends
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector stores
qdrant-client==1.7.0
weaviate-client==3.25.3

# Additional utilities
httpx==0.25.2
sqlite-fts4==1.0.3
python-multipart==0.0.6